Techdebt: Move some tests around
